backend: added an event-listener 'updateScores' to the server socket defined in 'index.js'
This commit adds an event-listener 'updateScores' to the socket defined in 'index.js', as well as tests for said event-listener in the file 'index.test.js'. The players' total scores are maintained seperately from two hidden scores (the players' artist and trickster scores), which are used for tracking the players' performance within the artist and trickster roles.

* added an event-listener 'updateScores' to 'index.js'
* added jest as a dependency in 'package.json'
* added a folder '_tests_' for storing test files
* added a file 'index.tests.js' for testing 'index.js'

addresses #14