feat:Made request handling/database querying
This commit includes new async functions for database querying for both categories and words.
Random categroy randomization is already made in the same querying function.
Responses for Get requests have been implemented to receive a seed for random number generation,
so that all players will get the same random numbers. For this reason an external package was
included for random generation with seeds (randomseed).

- Installed Node driver for MongoDB
- Installed randomseed package for random number generation with seeds
- get request handling was made for Categories, sends back 3 random catgeories (same for all players)
- function for retrieving words of a particular category was made
- included .env in .gitignore just in case