Testing: completed testing for fetching
The commmit includes the creationg of 2 new files, app.js and app.test.js.
App.js is the same as index.js but it contains a line at the end to allow
it to be exported for testing. Tests are build in app.test.js.
Supertest was installed to test requests.

These changes do not affect anybody elses code.