backend: add a variable 'roomGuesses' with listeners 'submitGuess' and 'changeGuesses' to communicate with the frontend
This commit creates an array 'roomGuesses' meant to mirror the behaviour of the 'roomUsers' array to share guesses between users on the frontend. It includes the event listeners 'submitGuess' which will add a new guess to the 'roomGuesses' array before sharing the new list fo guesses with the frontend as well as the 'changeGuesses' event listener which takes in a modified verios of the guess array based on the votes and updates the list on the backend before sending it to the frontend of all users.

* adds a variable 'roomGuesses' to Index.js
* adds the event listeners 'submitGuess' and 'changeGuesses' to Index.js

addresses COSC481W-2024Winter/Fictionary-Frontend#36